chore: update operation IDs to be consistent and follow programmatic conventions.
As per OpenAPI specification guidance for operationID field, follow programmatic
conventions for operationId field. To me that means as if I were defining a method
or function name.

This field is not used in the API docs as far as I can tell. There were
many different naming conventions across APIs. I've aligned to be camel case.

This help facilitate API SDK generation as all endpoints can have a function named
based on the operation ID and it have sufficient context to be unique amongst all APIs