[DEVDOCS-5864]/[ANA-4629] Users, Update AddUserToStore docs to reflect security patch  (#201)
<!-- Ticket number or summary of work -->
# [DEVDOCS-5864]
Jira: [ANA-4629](https://bigcommercecloud.atlassian.net/browse/ANA-4629)


## What changed?
* Updating the AddUserToStore documentation to reflect the security
patch that the IAM team tackled as part of ANA-4629. The vulnerability
there was that it was possible for attackers to enumerate through user
ID's and retrieve all the emails of users throughout the platform if
they were successful in using an ID that belongs to someone. The IAM
team recommends that we instead use email as the identifier for users as
part of this specific mutation, however we are still allowing for a set
of accounts that have existing Users API tokens to continue using the
original API contract as before.


## Release notes draft
<!-- Provide an entry for the release notes using simple, conversational
language. Don't be too technical. Explain how the change will benefit
the merchant and link to the feature.

Examples:
* The newly-released [X feature] is now available to use. Now, you’ll be
able to [perform Y action].
* We're happy to announce [X feature], which can help you [perform Y
action].
* [X feature] helps you to create [Y response] using the [Z query
parameter]. Now, you can deliver [ex, localized shopping experiences for
your customers].
* Fixed a bug in the [X endpoint]. Now the [Y field] will appear when
you click [Z option]. -->
* 

## Anything else?
<!-- Add related PRs, salient notes, additional ticket numbers, etc. -->

ping {names}


[ANA-4629]:

https://bigcommercecloud.atlassian.net/browse/ANA-4629?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ



[DEVDOCS-5864]:
https://bigcommercecloud.atlassian.net/browse/DEVDOCS-5864?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[ANA-4629]:
https://bigcommercecloud.atlassian.net/browse/ANA-4629?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[ANA-4629]:
https://bigcommercecloud.atlassian.net/browse/ANA-4629?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: Michael Angelo Lopez <47608798+mikelopez100@users.noreply.github.com>