DEVDOCS-6216 - Account Hierarchy + Content Refresh (#844)
With the GA of Account Hierarchies, several new endpoints are added. In
addition, this is the first update of the B2B Content Refresh.

<!-- Ticket number or summary of work -->
# [DEVDOCS-6216]


## What changed?
* Account Hierarchies adds four new endpoints for allowing parent/child
relationships between companies
* Company API information is fully refreshed with more information about
the structure and usage about each endpoint and the API as a whole.

## Release notes draft
* We're happy to announce the release of Account Hierarchies which
allows
    * Parent/Child relationships between companies
    * Company hierarchy trees
    * Inherited permission dependencies for the customer experience
* We've done a full refresh of the B2B Edition Company endpoints,
providing extended details about every endpoint, its parameters,
queries, and fields, and the relevant functionality of Companies as a
whole within B2B Edition.

ping {bc-sylvialin,bc-Vince,CNanninga}


[DEVDOCS-6216]:
https://bigcommercecloud.atlassian.net/browse/DEVDOCS-6216?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Traci Porter <traci.porter@bigcommerce.com>