chore(handlebars-reference.mdx): Clean up tables for increased readability and ease of navigation. (#896)
*External contributor*

Hi BC dev docs team,

I'm currently sending PR's to the stencil-handebars repo, and I here to
offer to contribute the associated documentation to the accepted helper
changes here. This is just a proposal PR opened to start getting used to
opening PR's in the way you would prefer over here for when I submit
offers for associated documentation. This proposal is not adding any
content per-se to the doc's, it's mostly dealing with a minor grief I've
had for quite some time regarding the density and sparseness of
information on the `handlebars helpers reference` section of the docs
and adding addendum information things like "what features of out of
handlebars 3 and 4 actually work in themes".

- Jordan

# What, Why?

This PR represents a proposal to change the layout of the Stencil
Helpers documentation page to allow for increased consolidation of
sparse information.

Proposed layout is as follows:
1. Introduction to handlebars documentation.
2. Standard features and deviations notes.
3. Grouped lists of BigCommerce specified / developed helpers.
4. Move third party helpers list from bottom of page where it is
"hidden" / out of sight.
5. Documentation of custom helpers.
6. Contribution notes and external resources.

## What was changed?
* Documentation for "Handlebars Helpers" has been updated to help with
being able to identify and parse information.

## Release notes draft
Adjusted the layout of Stencil Handlebars Reference documentation to
help theme developers navigate resources.

## Anything else?
Not in this PR no.

@jairo-bc