docs(orders): DEVDOCS-6426 Update date_shipped description (#994)
<!-- Ticket number or summary of work -->
# [DEVDOCS-6426]


## What changed?
Merchants found it ambiguous if date is for the first shipment of the
order or when the order is fully shipped.
Updated description to confirm date if for when the order is fully
shipped.

## Release notes draft
<!-- Provide an entry for the release notes using simple, conversational
language. Don't be too technical. Explain how the change will benefit
the merchant and link to the feature.

Examples:
* The newly-released [X feature] is now available to use. Now, you’ll be
able to [perform Y action].
* We're happy to announce [X feature], which can help you [perform Y
action].
* [X feature] helps you to create [Y response] using the [Z query
parameter]. Now, you can deliver [ex, localized shopping experiences for
your customers].
* Fixed a bug in the [X endpoint]. Now the [Y field] will appear when
you click [Z option]. -->
* 

## Anything else?
<!-- Add related PRs, salient notes, additional ticket numbers, etc. -->

ping @bigcommerce/team-orders @bigcommerce/team-shipping 


[DEVDOCS-6426]:
https://bigcommercecloud.atlassian.net/browse/DEVDOCS-6426?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: Terra Hyde <terra.hyde@bigcommerce.com>