docs(orders): ORDERS-7300 remove merchant_calculated_override from re… (#1128)
…fund_quote endpoint

<!-- Ticket number or summary of work -->
# [ORDERS-7300]


## What changed?
<!-- Provide a bulleted list in the present tense -->
* remove `merchant_calculated_override` from request payload of `POST
/refund_quote` endpoint

This is because we never support that in the first place, so this just
causes confusion.
(https://github.com/bigcommerce/bigcommerce/blob/master/app/Order/DomainModel/Refund/QuoteRequest.php)

`merchant_calculated_override` is only usable on the `POST /refunds`
endpoint when finally submit the refund.

## Release notes draft
<!-- Provide an entry for the release notes using simple, conversational
language. Don't be too technical. Explain how the change will benefit
the merchant and link to the feature.

n/a 

## Anything else?
<!-- Add related PRs, salient notes, additional ticket numbers, etc. -->

ping @bigcommerce/dev-docs @bigcommerce/team-orders 

[ORDERS-7300]:
https://bigcommercecloud.atlassian.net/browse/ORDERS-7300?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ