DEVDOCS-6561 - Clarify 'updated_at' field description (#1131)
Updated description for 'updated_at' field to clarify its current
behavior regarding API calls. This will be fixed at some point in the
future, but the workload is greater than the impact.

<!-- Ticket number or summary of work -->
# [DEVDOCS-6561]


## What changed?
* Clarified the current behaviour of the `updated_at` field for consent
management.

## Release notes draft
* Updated description for 'updated_at' field to clarify its current
behavior regarding API calls. This will be fixed at some point in the
future, but the workload is greater than the impact.
*
https://developer.bigcommerce.com/docs/rest-management/customers/consent

## Anything else?
<!-- Add related PRs, salient notes, additional ticket numbers, etc. -->

ping { @bigcommerce/dev-docs-team }


[DEVDOCS-6561]:
https://bigcommercecloud.atlassian.net/browse/DEVDOCS-6561?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ