Clarify Simulated Group
Adds a reference to more clearly define what the Simulated Group is and means to editing this component in Makeswift.

I know this makes things slightly redundant because you are told to select a Simulated Group before adding content for Group A just below this. However, both myself and a merchant's developer missed this the first time through and were very confused as to why we had Targeted Groups but content didn't appear for them, so I think a further definition alongside Simulated Group's definition is warranted. It's not really clear that this is a requirement to create the desired content, in my opinion, without that.