Clarify Simulated Group (#1200)
<!-- Ticket number or summary of work -->
## Summary
Adds a reference to more clearly define what the Simulated Group is and
means to editing this component in Makeswift.

I know this makes things slightly redundant because you are told to
select a Simulated Group before adding content for Group A just below
this. However, both myself and a merchant's developer missed this the
first time through and were very confused as to why we had Targeted
Groups but content didn't appear for them, so I think a further
definition alongside Simulated Group's definition is warranted. It's not
really clear that this is a requirement to create the desired content,
in my opinion, without that.


## What changed?
<!-- Provide a bulleted list in the present tense -->
* A note is added below the Simulated Group definition that clearly
defines why it must be selected.

## Release notes draft
<!-- Provide an entry for the release notes using simple, conversational
language. Don't be too technical. Explain how the change will benefit
the merchant and link to the feature.

Examples:
* The newly-released [X feature] is now available to use. Now, you’ll be
able to [perform Y action].
* We're happy to announce [X feature], which can help you [perform Y
action].
* [X feature] helps you to create [Y response] using the [Z query
parameter]. Now, you can deliver [ex, localized shopping experiences for
your customers].
* Fixed a bug in the [X endpoint]. Now the [Y field] will appear when
you click [Z option]. -->
* 
N/A
## Anything else?
<!-- Add related PRs, salient notes, additional ticket numbers, etc. -->
N/A